refactor(config-ui): move components/delete-button to components/action/delete-button
