+++ config-ui/src/components/action/delete-button/index.tsx
