+++ config-ui/src/components/action/index.ts
@@ -16,14 +16,4 @@
  *
  */
 
-export * from './loading';
-export * from './divider';
-export * from './page-header';
-export * from './selector';
-export * from './dialog';
-export * from './table';
-export * from './toast';
-export * from './logo';
-export * from './card';
-export * from './inspector';
-export * from './action';
+export * from './delete-button';
